Add standalone mode and Ignore fuzzing
We wanted to fix bug in coverage collection in qemu code,
but it resulted in we get very large coverage output (>1G).
Then, we want to separate standalone mode
to get full coverage collection in a execution
from fuzzing mode in which qemu emits compressed bitmap
instead of full coverage.
To achieve this, as first step, we add support for standalone mode
and temporalilly igmore fuzzing mode.
Fuzzing mode will be re-enabled in the next commit.